docs: Add eda_p2 file
This file contains Nicolas' take on the EDA of the given training data. It also includes the feature selection process using regularization, wherein the final models per regularization method will be used for comparing other models or even used for prediction should one of them be the most optimized models.